backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.31.0
httpx==0.26.0

langchain==0.1.16
langchain-community==0.0.36
openai==1.12.0
